Add support for external diagnostics. (#4)
Diagnostics to external files were not properly supported.  This adds
the changes necessary to fully support diagnostics which reference
files external to the buffer where the diagnostic was reported.

The first issue is that the position was not being computed correctly
since `flycheck-error-pos` uses the buffer in the diagnostic to
compute the position.  This is incorrect, when the diagnostic is for
an external file.  The correction was to make a copy of the diagnostic
and update the buffer within the diagnostic to use the buffer where
the problem resides, not where it was reported.  Once that is updated,
invoking Flycheck APIs which utilize the buffer in the diagnostic will
provide the correct position.  This is the same approach used in
`flycheck-jump-to-error`.

A second issue occurs due to the placement of the `consult--candidate`
and `consult--type` properties.  These were being placed at the
beginning of the filename string in the formatted diagnostic.  This is
problematic when multiple filenames of different lengths exist in the
diagnostics list, as additional padding may be added before the
filename for alignment purposes.  As a result, these properties were
not guaranteed to be at the first position of the formatted string, as
expected.  The correction was to apply the attributes after the
formatted string is computed.

The `consult--candidate` property was also updated with overlay
information to add full match highlighting.

Finally, the sort predicate (i.e., `consult-flycheck--sort-predicate`)
was updated to include the filename as part of the sorting criteria.